requirements.txt
numpy==1.21.3
opencv_python==4.5.4.60
Pillow==9.2.0
requests==2.26.0
